[refactor] Parser에서 Caculator를 생성하도록 수정
