+++ src/main/java/controller/Calculator.java
@@ -13,37 +13,12 @@ public class Calculator {
     private final List<Number> numbers;
     private final List<Operator> operators;
 
-    public Calculator(String expression) {
-        this(Parser.parse(expression));
-    }
-
-    public Calculator(String[] expression) {
-        this(parseNumbers(expression), parseOperators(expression));
-    }
-
     public Calculator(List<Number> numbers, List<Operator> operators) {
         this.numbers = numbers;
         this.operators = operators;
     }
 
-    private static List<Number> parseNumbers(String[] expression) {
-        List<Number> numbers = new ArrayList<>();
 
-        for (int i = 0; i < expression.length; i+=2) {
-            numbers.add(new domain.Number(expression[i]));
-        }
-
-        return numbers;
-    }
-
-    private static List<Operator> parseOperators(String[] expression) {
-        List<Operator> operators = new ArrayList<>();
-        for (int i = 1; i < expression.length - 1; i+=2) {
-            operators.add(OperatorFactory.getOperator(expression[i]));
-        }
-
-        return operators;
-    }
 
     public Number calculate() {
         Number result = numbers.getFirst();
